mcp/pyproject.toml
[project]
name = "fastmcp-playground"
version = "0.1.0"
description = "A playground to test and develop an MCP server with FastMCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.0.2",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "llm-sandbox[docker]>=0.3.24",
]

[dependency-groups]
dev = [
    "pandas-stubs>=2.3.2.250926",
    "ruff>=0.14.4",
]
